+++ zcash_client_sqlite/src/lib.rs
@@ -259,6 +259,37 @@ impl<C: Borrow<rusqlite::Connection>, P: consensus::Parameters> WalletRead for W
     type Error = SqliteClientError;
     type AccountId = AccountId;
 
+    fn validate_seed(
+        &self,
+        account_id: Self::AccountId,
+        seed: &SecretVec<u8>,
+    ) -> Result<bool, Self::Error> {
+        if let Some(account) = self.get_account(account_id)? {
+            if let Account::Zip32(HDSeedAccount(_, account_index, ufvk)) = account {
+                let usk = UnifiedSpendingKey::from_seed(
+                    &self.params,
+                    &seed.expose_secret()[..],
+                    account_index,
+                )
+                .map_err(|_| SqliteClientError::KeyDerivationError(account_index))?;
+
+                // Keys are not comparable with `Eq`, but addresses are, so we derive what should
+                // be equivalent addresses for each key and use those to check for key equality.
+                UnifiedAddressRequest::all().map_or(Ok(false), |ua_request| {
+                    Ok(usk
+                        .to_unified_full_viewing_key()
+                        .default_address(ua_request)
+                        == ufvk.default_address(ua_request))
+                })
+            } else {
+                Err(SqliteClientError::UnknownZip32Derivation)
+            }
+        } else {
+            // Missing account is documented to return false.
+            Ok(false)
+        }
+    }
+
     fn get_account(&self, account_id: Self::AccountId) -> Result<Option<Account>, Self::Error> {
         wallet::get_account(self.conn.borrow(), &self.params, account_id)
     }
@@ -1225,9 +1256,10 @@ extern crate assert_matches;
 
 #[cfg(test)]
 mod tests {
+    use secrecy::SecretVec;
     use zcash_client_backend::data_api::{AccountBirthday, WalletRead, WalletWrite};
 
-    use crate::{testing::TestBuilder, DEFAULT_UA_REQUEST};
+    use crate::{testing::TestBuilder, AccountId, DEFAULT_UA_REQUEST};
 
     #[cfg(feature = "unstable")]
     use {
@@ -1237,6 +1269,36 @@ mod tests {
             consensus::Parameters, transaction::components::amount::NonNegativeAmount,
         },
     };
+
+    #[test]
+    fn validate_seed() {
+        let st = TestBuilder::new()
+            .with_test_account(AccountBirthday::from_sapling_activation)
+            .build();
+
+        assert!({
+            let account = st.test_account().unwrap().0;
+            st.wallet()
+                .validate_seed(account, st.test_seed().unwrap())
+                .unwrap()
+        });
+
+        // check that passing an invalid account results in a failure
+        assert!({
+            let account = AccountId(3);
+            !st.wallet()
+                .validate_seed(account, st.test_seed().unwrap())
+                .unwrap()
+        });
+
+        // check that passing an invalid seed results in a failure
+        assert!({
+            let account = st.test_account().unwrap().0;
+            !st.wallet()
+                .validate_seed(account, &SecretVec::new(vec![1u8; 32]))
+                .unwrap()
+        });
+    }
 
     #[test]
     pub(crate) fn get_next_available_address() {